+++ projects/client/audio/aura_sequencing/source/aura/sequencing/Conductor.cpp
@@ -8,7 +8,7 @@ namespace aura {
 //    set_chord(get_current_chord_structure().get_chords()[0]);
     }
 
-    void Conductor::set_chord(const Chord_Instance chord, float offset, float start, float end) {
+    void Conductor::set_chord(const Chord_Instance chord, float offset, double start, double end) {
       this->chord = chord;
       if (recorder)
         recorder->add_event(new Chord_Event(chord, offset, start, end));
